requirements.txt
phidata
feedparser
newspaper3k
requests
selenium
python-dotenv
google-generativeai